Split methods related to confchanges into separate file
